test: move inheritance_break into test_branches
